server/pyproject.toml
[project]
name = "zener-server"
version = "0.1.0"
description = "AI Remote Assistance Platform - Backend"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-cloud-firestore>=2.16.0",
    "firebase-admin>=6.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py312"
